Rename: 공통 컴포넌트 components/common/ 에서 components/ 로 경로 수정
