+++ src/components/LogText/HintIcon/index.tsx
