+++ src/components/LogText/index.tsx
